+++ src/saivenky/neural/IOutputLayer.java
@@ -4,8 +4,8 @@ package saivenky.neural;
  * Created by saivenky on 2/22/17.
  */
 public interface IOutputLayer extends ILayer {
-    void setExpected(double[][] expected);
-    void setSignalCostGradient(double[][] cost);
-    void getPredicted(double[][] predicted);
+    void setExpected(float[][] expected);
+    void setSignalCostGradient(float[][] cost);
+    void getPredicted(float[][] predicted);
     int size();
 }
