+++ src/saivenky/neural/c/MaxPoolingLayer.java
@@ -22,6 +22,6 @@ public class MaxPoolingLayer extends Layer {
     private native long destroy(long nativeLayerPtr);
 
     @Override
-    public void gradientDescent(double rate) {
+    public void gradientDescent(float rate) {
     }
 }
